fix(mock-structure-handoff): implement all 4 phases of fix plan
Phase 1: Route mock_structure through spec data flow
- Drop bypass mockStructureDeclaration field from AutofixImplementerOutput
- Stop filtering mock_structure declarations in parse()
- Remove bypass branch from implementer's extractApplied
- Revert testWriter appliesTo to spec form (fixTarget === 'test' only)
- Collapse testWriter buildInput to single mock-restructure branch
- Clear pendingMockStructureHandoffs one-shot per spec US-004 AC#3
- Update integration test mocks to use testEditDeclarations

Phase 2: Add defensive empty-files guard
- Add validation for empty FILES in validateMockStructureFiles
- Mark empty declarations as invalid with no files/nonTest
- Add unit test for empty files defensive partition

Phase 3: Pass packageDir to resolveTestFilePatterns
- Thread packageDir through validate() call site
- Update runIsolationGuard signature to accept optional packageDir
- Update guard call site in cycle to pass packageDir
- Add unit test verifying packageDir forwarding

Phase 4: Test file consolidation
- Move config-validation tests from autofix-guards.test.ts to autofix-guards-spec.test.ts
- Delete placeholder file (now autofix-guards.test.ts)
- Rename autofix-guards-spec.test.ts to autofix-guards.test.ts
- All 4 config tests now in one file per naming convention

Fixes: All core integration tests pass, spec conformance restored
